+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/MapboxNavigationViewFragmentActivity.kt
@@ -2,6 +2,7 @@ package com.mapbox.navigation.qa_test_app.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.WindowCompat
 import com.mapbox.navigation.qa_test_app.databinding.LayoutActivityNavigationViewFragmentBinding
 
 class MapboxNavigationViewFragmentActivity : AppCompatActivity() {
@@ -12,5 +13,7 @@ class MapboxNavigationViewFragmentActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = LayoutActivityNavigationViewFragmentBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        WindowCompat.setDecorFitsSystemWindows(window, false)
     }
 }